The filter_telnum prefix column is int(32) preventing filtering for any telephone numbers higher than 4294967295. Changing the column to bigint allows filters to be created from the web interface but they are ignored by the voipmonitor sniffer. - make it bingint in db and chnage unsigned int to unsigned long long supporting numbers 0 to 18,446,744,073,709,551,615

+++ cdrtable.sql
@@ -14,7 +14,7 @@ CREATE TABLE `filter_ip` (
 
 CREATE TABLE `filter_telnum` (
   `id` int(32) NOT NULL auto_increment,
-  `prefix` int(32) unsigned default NULL,
+  `prefix` bigint(32) unsigned default NULL,
   `fixed_len` int(32) unsigned default 0,
   `direction` tinyint(8) default 0,
   `rtp` tinyint(1) default '0',
